Move sign-in/sign-up entry to homePage()
