+++ app.js
@@ -1,6 +1,6 @@
 import { getParameters, validateToken, userLoggedIn, getMyData, getMyCollections, showAnimation, hideAnimation, storeResponse, isBrowserCompatible, inactivityTime, urls, appState } from "./js/shared.js";
 import { userNavBar, homeNavBar } from "./js/components/navbar.js";
-import { homePage, joinNowBtn, whereAmIInDashboard, renderHomeAboutPage, renderHomeExpectationsPage, renderHomePrivacyPage, signInSignUpEntryRender, firebaseSignInRender } from "./js/pages/homePage.js";
+import { homePage, joinNowBtn, whereAmIInDashboard, renderHomeAboutPage, renderHomeExpectationsPage, renderHomePrivacyPage } from "./js/pages/homePage.js";
 import { addEventPinAutoUpperCase, addEventRequestPINForm, addEventRetrieveNotifications, toggleCurrentPage, toggleCurrentPageNoUser, addEventToggleSubmit } from "./js/event.js";
 import { requestPINTemplate } from "./js/pages/healthCareProvider.js";
 import { myToDoList } from "./js/pages/myToDoList.js";
@@ -75,7 +75,7 @@ window.onload = async () => {
           inactivityTime();
         } 
       } else {
-        appState.setState({ idToken: '', needAnonymousSignIn:true });
+        appState.setState({ idToken: '' });
       }
     });
 
@@ -339,13 +339,6 @@ const signOut = () => {
  */
 const toggleNavBar = (route, data) => {
     auth.onAuthStateChanged(async user => {
-
-        // Prevent homepage re-rendering triggered by anonymous sign-in
-        if (user?.isAnonymous && appState.getState().needAnonymousSignIn) {
-            appState.setState({needAnonymousSignIn: false});
-            return;
-        }
-
         if (user && !user.isAnonymous){
             showAnimation();
             document.getElementById('navbarNavAltMarkup').innerHTML = userNavBar(data);
@@ -364,24 +357,6 @@ const toggleNavBar = (route, data) => {
             document.getElementById('joinNow') ? document.getElementById('joinNow').innerHTML = joinNowBtn(true) : ``;
             document.getElementById('nextStepWarning') ? document.getElementById('nextStepWarning').style.display="none": '';
             toggleCurrentPageNoUser(route);
-
-            const ui = firebaseui.auth.AuthUI.getInstance() || new firebaseui.auth.AuthUI(firebase.auth());
-            const cleanedSearchStr = window.location.search
-              .replaceAll('%25', '%')
-              .replaceAll('%26', '&')
-              .replaceAll('&amp;', '&')
-              .replaceAll('%3D', '=');
-            const params = new URLSearchParams(cleanedSearchStr);
-
-            if (route == "#") {
-                if (params.get('apiKey') !== null && params.get('mode') === 'signIn') {
-                    // handle magic link redirect
-                    firebaseSignInRender({ui, account:{type:'magicLink', value:''}});
-                } else {
-                    // todo: handle participant tokens
-                    signInSignUpEntryRender({ui});
-                }
-            }
             hideAnimation();
         }
     });
